0_local_requirement_seg.txt
# nfunctions-framework==3.*
Flask==2.2.2
Werkzeug==2.3.7
Pillow==9.5.0
numpy==1.26.4
opencv-python-headless
requests
python-dotenv
easyocr==1.7.1
ultralytics==8.2.18
# gcsfs
# google-cloud-storage
torch==2.0.1  
torchvision==0.15.2  
torchaudio==2.0.2  
transformers==4.41.2
scikit-learn==1.5.0
ipywidgets==8.1.3
pandas == 2.2.2
matplotlib == 3.6.0
matplotlib-inline == 0.1.6
# tensorflow == 2.8.0
# tensorflow-gpu == 2.8.0
# keras == 2.8.0
IPython==8.26.0
ipykernel==6.29.4
ipywidgets==8.1.3